// backend/src/interfaces/report.interface.ts
export enum Schet {
  S00 = "S00", // СЧЕТА ДЛЯ ВВОДА ОСТАТКОВ И ЗАКРЫТИЯ ЗП
  S01 = "S01", // СЧЕТА УЧЕТА ОСНОВНЫХ СРЕДСТВ
  S02 = "S02", // НАКОПЛЕННАЯ АМОРТИЗАЦИЯ ОС
  S10 = "S10", // СЧЕТА УЧЕТА МАТЕРИАЛОВ
  S11 = "S11", // СЧЕТА УЧЕТА ИНСТРУМЕНТОВ НА СКЛАДЕ
  S12 = "S12", // СЧЕТА УЧЕТА ИНСТРУМЕНТОВ У КЛИЕНТА (В АРЕНДЕ)
  S13 = "S13", // СЧЕТА УЧЕТА ИНСТРУМЕНТОВ СУБАРЕНДЫ (У ПАРТНЁРА / ВЫДАНО КЛИЕНТУ)
  S20 = "S20", // СЧЕТА УЧЕТА ОСНОВНОГО ПРОИЗВОДСТВА И СЧЕТА УЧЕТА РАСХОДОВ ПЕРИОДА
  S21 = "S21", // СЧЕТА УЧЕТА ПОЛУФАБРИКАТОВ СОБСТВЕННОГО ПРОИЗВОДСТВА
  S23 = "S23", // СЧЕТ УЧЕТА ВСПОМОГАТЕЛЬНОГО ПРОИЗВОДСТВА
  S28 = "S28", // СЧЕТА УЧЕТА ГОТОВОЙ ПРОДУКЦИИ
  S29 = "S29", // СЧЕТА УЧЕТА ТОВАРОВ
  S40 = "S40", // СЧЕТА К ПОЛУЧЕНИЮ ОТ КЛИЕНТОВ
  S41 = "S41", // СЧЕТА К ПОЛУЧЕНИЮ ОТ ЗАКАЗЧИКОВ
  S50 = "S50", // СЧЕТА УЧЕТА ДЕНЕЖНЫХ СРЕДСТВ В КАССЕ
  S51 = "S51", // СЧЕТА УЧЕТА ДЕНЕЖНЫХ СРЕДСТВ НА РАСЧЕТНОМ СЧЕТЕ
  S60 = "S60", // СЧЕТА К ОПЛАТЕ ПОСТАВЩИКАМ И ПОДРЯДЧИКАМ
  S64 = "S64", // СЧЕТА УЧЕТА РАСЧЕТОВ С ДОСТАВЩИКАМИ
  S66 = "S66", // СЧЕТА К ПОЛУЧЕНИЕ ОТ УЧРИДИТЕЛEЙ
  S67 = "S67", // СЧЕТА УЧЕТА ЗАРОБОТНОЙ ПЛАТЫ СОТРУДНИКОВ
  S71 = "S71", // СЧЕТА УЧЕТА ПОДОТЧЁТНЫХ СУММ (ЖАВОБГАР ШАХС)
  S65 = "S65", // СЧЕТА УЧЕТА БОНУСОВ ПОСРЕДНИКОВ (ВОСИТАЧИЛАР)
  S68 = "S68", // КОШЕЛКИ УЧРИДИТЕЛЕЙ,
  S90 = "S90", // СЧЕТА УЧЕТА ДОХОДОВ
  S91 = "S91", // СЧЕТА УЧЕТА РАСХОДОВ
  S93 = "S93", // ПРОЧИЕ ДОХОДЫ (БОШҚА ДАРОМАД)
}